updated user specific routes t check for session data

+++ flask06.py
@@ -39,8 +39,10 @@ with app.app_context():
 @app.route('/index')
 def index():
     #a_user = {'name': 'Lucas', 'email': "ldickso4@uncc.edu"}
-    a_user =  db.session.query(User).filter_by(email='ldickso4@uncc.edu').one()
-    return render_template('index.html', user = a_user)
+    if session.get('user'):
+        return render_template('index.html', user = session['user'])
+    return render_template("index.html")
+    
 
 @app.route('/notes')
 def get_notes():
@@ -61,47 +63,54 @@ def get_note(note_id):
 @app.route('/notes/new', methods=['GET', 'POST'])
 def new_note():
     print('request method is: ', request.method)
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        from datetime import date
-        today = date.today()
-        today = today.strftime("%m-%d-%Y")
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            from datetime import date
+            today = date.today()
+            today = today.strftime("%m-%d-%Y")
 
-        newEntry = Note(title, text, today)
-        db.session.add(newEntry)
-        db.session.commit()
+            newEntry = Note(title, text, today, session['user_id'])
+            db.session.add(newEntry)
+            db.session.commit()
 
-        return redirect(url_for('get_notes'))
+            return redirect(url_for('get_notes'))
+        else: 
+            return render_template('new.html', user=session['user'])
     else: 
-        a_user = db.session.query(User).filter_by(email='ldickso4@uncc.edu').one()
-        return render_template('new.html', user=a_user)
+        return render_template(url_for('login'))
 
 @app.route('/notes/edit/<note_id>', methods=['GET', 'POST'])
 def update_note(note_id):
-    if request.method == 'POST':
-        title = request.form['title']
-        text = request.form['noteText']
-        note = db.session.query(Note).filter_by(id=note_id).one()
-        note.title = title
-        note.text = text
-        db.session.add(note)
+    if session.get('user'):
+        if request.method == 'POST':
+            title = request.form['title']
+            text = request.form['noteText']
+            note = db.session.query(Note).filter_by(id=note_id).one()
+            note.title = title
+            note.text = text
+            db.session.add(note)
+            db.session.commit()
+
+            return redirect(url_for('get_notes'))
+        else:
+            my_note = db.session.query(Note).filter_by(id=note_id).one()
+
+            return render_template('new.html', note=my_note, user=session['user'])
+    else:
+        return redirect(url_for('login'))
+
+@app.route('/notes/delete/<note_id>', methods=['POST'])
+def delete_note(note_id):
+    if session.get('user'):
+        my_note = db.session.query(Note).filter_by(id=note_id).one()
+        db.session.delete(my_note)
         db.session.commit()
 
         return redirect(url_for('get_notes'))
     else:
-        a_user = db.session.query(User).filter_by(email='ldickso4@uncc.edu').one()
-        my_note = db.session.query(Note).filter_by(id=note_id).one()
-
-        return render_template('new.html', note=my_note, user=a_user)
-
-@app.route('/notes/delete/<note_id>', methods=['POST'])
-def delete_note(note_id):
-    my_note = db.session.query(Note).filter_by(id=note_id).one()
-    db.session.delete(my_note)
-    db.session.commit()
-
-    return redirect(url_for('get_notes'))
+        return redirect(url_for('login'))
 
 @app.route('/register', methods=['POST', 'GET'])
 def register():
@@ -143,7 +152,7 @@ def login():
     else:
         # form did not validate or GET request
         return render_template("login.html", form=login_form)
-        
+
 
 app.run(host=os.getenv('IP', '127.0.0.1'),port=int(os.getenv('PORT', 5000)),debug=True)
